chore(bg-editor): rename to WallpaperEditor more
