+++ src/containers/layout/GlobalLayout/index.tsx
@@ -24,7 +24,7 @@ import Header from '@/widgets/Header'
 
 import type { TStore } from './store'
 import SEO from './SEO'
-import CustomBg from './CustomBg'
+import Wallpaper from './Wallpaper'
 
 import { CustomScroller, Footer, ModeLine } from './dynamic'
 
@@ -60,7 +60,7 @@ const GlobalLayoutContainer: FC<TProps> = ({
 
   return (
     <ThemePalette>
-      <CustomBg />
+      <Wallpaper />
       <Wrapper>
         <SEO metric={metric} config={seoConfig} />
         <InnerWrapper metric={metric} sidebarPin={sidebarPin}>
