+++ src/containers/layout/GlobalLayout/styles/wallpaper.ts
